requirements.txt
pandas>=2.0.0
numpy>=1.24.0
yfinance>=0.2.40
requests
feedparser
beautifulsoup4
vaderSentiment
pandas
python-dateutil
lxml               # speeds up HTML parsing
newspaper3k        # article text extraction
tqdm               # progress bars for batch scraping
schedule           # for daily updates
pyarrow # for parquet storage dependency
openai # I'm actually not sure if this is how it works but I installed it to help on my command line
alpaca-trade-api
python-dotenv # for .env file for secret keys
alpaca-py
psutil
pytest # for testing
# pipx install aider-ce
urllib3>=1.26.18
dash>=2.13.0
plotly>=5.17.0
